requirements.txt
langchain>=0.3.0
langchain-google-vertexai>=2.0.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.40.0
rich>=13.0.0
